feat(dts): introduce the `dtsMode` to control the behavior of generating dts file
